Move stopwaiter to its own go package
