+++ arbnode/inbox_reader.go
@@ -15,8 +15,8 @@ import (
 	flag "github.com/spf13/pflag"
 
 	"github.com/offchainlabs/nitro/arbutil"
-	"github.com/offchainlabs/nitro/util"
 	"github.com/offchainlabs/nitro/util/arbmath"
+	"github.com/offchainlabs/nitro/util/stopwaiter"
 )
 
 type InboxReaderConfig struct {
@@ -44,7 +44,7 @@ var TestInboxReaderConfig = InboxReaderConfig{
 }
 
 type InboxReader struct {
-	util.StopWaiter
+	stopwaiter.StopWaiter
 
 	// Only in run thread
 	caughtUp          bool
